backend/pyproject.toml
[project]
name = "sku-ops-backend"
version = "0.1.0"
description = "SKU Operations — backend API"
authors = [{ name = "nooz" }]
requires-python = ">=3.13,<3.14"
readme = "README.md"
license = "MIT"

dependencies = [
    # ── Core ──
    "fastapi>=0.110.1",
    "uvicorn>=0.25.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.6.4",
    "email-validator>=2.2.0",
    "python-multipart>=0.0.9",
    # ── Auth ──
    "pyjwt>=2.10.1",
    "bcrypt==4.1.3",
    "passlib>=1.7.4",
    # ── Database ──
    "aiosqlite>=0.19.0",
    "asyncpg>=0.29.0",
    # ── AI / LLM ──
    "pydantic-ai>=1.0.0",
    "pydantic-ai-slim[anthropic]>=1.0.0",
    "openai>=1.40.0",
    "tiktoken>=0.7.0",
    "rank-bm25>=0.2.2",
    "numpy>=1.26.0",
    # ── Document processing ──
    "Pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "pypdf>=4.0.0",
    "pdf2image>=1.17.0",
    # ── HTTP / OAuth ──
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "requests-oauthlib>=2.0.0",
    "cryptography>=42.0.8",
    # ── Config / Serialization ──
    "PyYAML>=6.0",
    "tzdata>=2024.2",
    # ── Production hardening ──
    "python-json-logger>=2.0.7",
    "prometheus-client>=0.20.0",
    "sentry-sdk[fastapi]>=1.40.0",
    "redis>=7.3.0",
]
